app ya con unidades

+++ app/src/main/java/com/cesar/bocana/data/model/Product.kt
@@ -4,31 +4,26 @@ import com.google.firebase.firestore.DocumentId
 import com.google.firebase.firestore.ServerTimestamp
 import java.util.Date
 
-
 data class Product(
-    @DocumentId // Mapea el ID del documento Firestore a este campo
+    @DocumentId
     val id: String = "",
-
     val name: String = "",
-    val unit: String = "", // Unidad de medida (Kg, Bolsas, Cajas, Piezas)
-    val minStock: Long = 0L, // Stock mínimo permitido (usamos Long para cantidades)
-    val providerDetails: String = "", // Detalles/Proveedor
-
-    // Stock actual en cada almacén
-    val stockMatriz: Long = 0L,
-    val stockCongelador04: Long = 0L,
-    val totalStock: Long = 0L, // Campo calculado (stockMatriz + stockCongelador04)
-
-    @ServerTimestamp // Firestore asignará la fecha del servidor al crear
+    val unit: String = "",
+    val minStock: Double = 0.0,
+    val providerDetails: String = "",
+    val stockMatriz: Double = 0.0,
+    val stockCongelador04: Double = 0.0,
+    val totalStock: Double = 0.0,
+    @ServerTimestamp
     val createdAt: Date? = null,
-    @ServerTimestamp // Firestore asignará/actualizará la fecha del servidor al modificar
+    @ServerTimestamp
     val updatedAt: Date? = null,
-    val lastUpdatedByName: String? = null, // <-- AÑADIDO
-       val isActive: Boolean = true // Opcional: Para "archivar" en lugar de borrar
+    val lastUpdatedByName: String? = null,
+    val isActive: Boolean = true
 ) {
     constructor() : this(
-        id = "", name = "", unit = "", minStock = 0L, providerDetails = "",
-        stockMatriz = 0L, stockCongelador04 = 0L, totalStock = 0L,
-        createdAt = null, updatedAt = null, lastUpdatedByName = null, isActive = true // <-- AÑADIDO EN CONSTRUCTOR
+        id = "", name = "", unit = "", minStock = 0.0, providerDetails = "",
+        stockMatriz = 0.0, stockCongelador04 = 0.0, totalStock = 0.0,
+        createdAt = null, updatedAt = null, lastUpdatedByName = null, isActive = true
     )
 }
